Adds AmortizingFixedRatePool and AmortizingFixedRateLoan classes.
These classes mimic the functionality provided in the experimental
AmortizingFixedRateBond class by adding the ability to pass in
a redemption argument which adjusts the principal received by
looking at the periodic redemption vector rather than the the
difference in notionals in the AmortizingFixedRateBond class.

The additions are made in the instruments for loans and also
in the pricingengines.

Changes and additions to configure.ac as well as the directory
specific Makefile.am files.